requirements.txt
pymysql
redis
flask